+++ __mocks__/@simonwep/pickr.js
@@ -1,0 +1,14 @@
+
+const noop = () => undefined
+exports.default = {
+  create: jest.fn((options) => {
+    const self = {
+      setColor: () => true,
+      on: () => self,
+      destroyAndRemove: noop,
+      options,
+    }
+
+    return self
+  }),
+}
